chore: moved 'achievementsByTpoic' function to 'helpers' folder from 'middleware' folder
